server/requirements.txt
Django==4.2.1
django-debug-toolbar==4.1.0
django-filter==23.2
django-tinymce==3.6.1
djangorestframework==3.14.0
model_bakery==1.11.0
Pillow==9.5.0
psycopg[binary]==3.1.9